pyproject.toml
[tool.poetry]
name = "flask-oidc"
version = "0.1.0"
description = ""
authors = ["showchan33 <showchan3333@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
Flask = "^3.0.3"
requests = "^2.32.3"
oauthlib = "^3.2.2"
python-dotenv = "^1.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
